pyproject.toml
[project]
name = "aegisforge"
version = "0.1.0"
description = "Enterprise autonomy agent for software delivery and automation"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "celery[redis]>=5.4.0",
    "redis>=5.2.0",
    "httpx>=0.28.0",
    "tenacity>=9.0.0",
    "pybreaker>=1.2.0",
    "structlog>=24.4.0",
    "prometheus-client>=0.21.0",
    "python-jose[cryptography]>=3.3.0",
    "authlib>=1.4.0",
    "pydantic-extra-types>=2.10.0",
    # LLM — Open-source models via OpenAI-compatible API (vLLM, Ollama, Together, Groq)
    "openai>=1.58.0",
    "tiktoken>=0.8.0",
    # RAG — vector search, open-source embeddings, chunking, reranking
    "pgvector>=0.3.6",
    "sentence-transformers>=3.3.0",
    "langchain-text-splitters>=0.3.4",
    "numpy>=2.1.0",
    "torch>=2.5.0",
    "transformers>=4.47.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx",
    "respx>=0.22.0",
    "testcontainers[postgres,redis]>=4.8.0",
    "hypothesis>=6.115.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "bandit>=1.8.0",
    "safety>=3.2.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aegisforge"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "S", "B", "A", "C4", "DTZ", "T20", "ICN", "PIE", "PT", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "PTH", "ERA"]
ignore = ["S101"]  # allow assert in tests

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=src/aegisforge --cov-report=term-missing --cov-fail-under=90"